Separated json contexts to ensure proper options
